+++ electron/ipc/exportIpc.ts
@@ -1,8 +1,9 @@
 import { ipcMain } from 'electron'
 import { ExportService } from '../services/exportService'
+import { secure } from '../utils/authGuard'
 
 export function registerExportIpc() {
-  ipcMain.handle('export:arpTable', () => ExportService.exportARPTable())
-  ipcMain.handle('export:changes', (_e, unacknowledgedOnly?: boolean) => ExportService.exportChanges(unacknowledgedOnly))
-  ipcMain.handle('export:networkUsage', (_e, networkId?: number) => ExportService.exportNetworkUsage(networkId))
+  ipcMain.handle('export:arpTable', secure(() => ExportService.exportARPTable()))
+  ipcMain.handle('export:changes', secure((_e, unacknowledgedOnly?: boolean) => ExportService.exportChanges(unacknowledgedOnly)))
+  ipcMain.handle('export:networkUsage', secure((_e, networkId?: number) => ExportService.exportNetworkUsage(networkId)))
 }